CPLP-1212 rename service to services
