+++ docker/Dockerfile-services-service
@@ -7,13 +7,13 @@ FROM mcr.microsoft.com/dotnet/sdk:6.0-alpine AS build
 WORKDIR /
 COPY src/ src/
 COPY keycloak/ keycloak/
-WORKDIR "/src/services/CatenaX.NetworkServices.Service.Service"
-RUN dotnet build "CatenaX.NetworkServices.Service.Service.csproj" -c Release -o /app/build
+WORKDIR "/src/services/CatenaX.NetworkServices.Services.Service"
+RUN dotnet build "CatenaX.NetworkServices.Services.Service.csproj" -c Release -o /app/build
 
 FROM build AS publish
-RUN dotnet publish "CatenaX.NetworkServices.Service.Service.csproj" -c Release -o /app/publish
+RUN dotnet publish "CatenaX.NetworkServices.Services.Service.csproj" -c Release -o /app/publish
 
 FROM base AS final
 WORKDIR /app
 COPY --from=publish /app/publish .
-ENTRYPOINT ["dotnet", "CatenaX.NetworkServices.Service.Service.dll"]
+ENTRYPOINT ["dotnet", "CatenaX.NetworkServices.Services.Service.dll"]
