+++ src/services/CatenaX.NetworkServices.Services.Service/Properties/launchSettings.json
