+++ src/services/CatenaX.NetworkServices.Services.Service/appsettings.json
@@ -34,7 +34,7 @@
       "ValidAudience": ""
     }
   },
-  "Service": {
+  "Services": {
     "ApplicationsMaxPageSize": 20
   }
 }
